+++ projects/archiver/main.ts
@@ -2,7 +2,7 @@ require('dotenv').config()
 export { handler as invoke } from './src/invoke'
 export { handler as issue } from './src/tasks/issue'
 export { handler as front } from './src/tasks/front'
-export { handler as image } from './src/tasks/image'
+export { handler as image } from './src/tasks/frontAndImage'
 export { handler as upload } from './src/tasks/upload'
 export { handler as zip } from './src/tasks/zip'
 export { handler as indexer } from './src/tasks/indexer'
