+++ projects/aws/lib/step-function.ts
@@ -1,88 +1,9 @@
-import cdk = require('@aws-cdk/core')
-import lambda = require('@aws-cdk/aws-lambda')
-import { Code, FunctionProps } from '@aws-cdk/aws-lambda'
-import s3 = require('@aws-cdk/aws-s3')
-import iam = require('@aws-cdk/aws-iam')
-import sfn = require('@aws-cdk/aws-stepfunctions')
-import tasks = require('@aws-cdk/aws-stepfunctions-tasks')
-import { toTitleCase } from './tools'
-import { Duration, Tag } from '@aws-cdk/core'
+import * as cdk from '@aws-cdk/core'
 import { Condition } from '@aws-cdk/aws-stepfunctions'
-interface StepFunctionProps {
-    stack: string
-    stage: string
-    deployBucket: s3.IBucket
-    outputBucket: s3.IBucket
-    backendURL: string
-    frontsTopicArn: string
-    frontsTopicRoleArn: string
-    guNotifyServiceApiKey: string
-}
-
-//Make sure you add the lambda name in riff-raff.yaml
-const taskLambda = (
-    name: string,
-    {
-        scope,
-        stack,
-        stage,
-        deployBucket,
-        outputBucket,
-        frontsTopicArn,
-        frontsTopicRole,
-    }: {
-        scope: cdk.Construct
-        stack: string
-        stage: string
-        deployBucket: s3.IBucket
-        outputBucket: s3.IBucket
-        frontsTopicArn: string
-        frontsTopicRole: iam.IRole
-    },
-    environment?: { [key: string]: string },
-    overrides?: Partial<FunctionProps>,
-) => {
-    const fn = new lambda.Function(
-        scope,
-        `EditionsArchiver${toTitleCase(name)}`,
-        {
-            functionName: `editions-archiver-stepmachine-${name}-${stage}`,
-            runtime: lambda.Runtime.NODEJS_10_X,
-            timeout: Duration.minutes(5),
-            memorySize: 1500,
-            code: Code.bucket(
-                deployBucket,
-                `${stack}/${stage}/archiver/archiver.zip`,
-            ),
-            handler: `index.${name}`,
-            environment: {
-                ...environment,
-                stage: stage,
-                bucket: outputBucket.bucketName,
-                topic: frontsTopicArn,
-                role: frontsTopicRole.roleArn,
-            },
-            initialPolicy: [
-                new iam.PolicyStatement({
-                    actions: ['*'],
-                    resources: [
-                        outputBucket.arnForObjects('*'),
-                        outputBucket.bucketArn,
-                    ],
-                }),
-                new iam.PolicyStatement({
-                    actions: ['sts:AssumeRole'],
-                    resources: [frontsTopicRole.roleArn],
-                }),
-            ],
-            ...overrides,
-        },
-    )
-    Tag.add(fn, 'App', `editions-archiver-${name}`)
-    Tag.add(fn, 'Stage', stage)
-    Tag.add(fn, 'Stack', stack)
-    return fn
-}
+import { Duration } from '@aws-cdk/core'
+import { StepFunctionProps, task } from './constructs'
+import * as iam from '@aws-cdk/aws-iam'
+import * as sfn from '@aws-cdk/aws-stepfunctions'
 
 export const archiverStepFunction = (
     scope: cdk.Construct,
@@ -104,7 +25,6 @@ export const archiverStepFunction = (
     )
 
     const lambdaParams = {
-        scope,
         stack,
         stage,
         deployBucket,
@@ -113,56 +33,44 @@ export const archiverStepFunction = (
         frontsTopicRole,
     }
     //Archiver step function
-    const issue = taskLambda('issue', lambdaParams, { backend: backendURL })
-
-    const issueTask = new sfn.Task(scope, 'Fetch Issue', {
-        task: new tasks.InvokeFunction(issue),
+    const issue = task(scope, 'issue', 'Fetch Issue', lambdaParams, {
+        backend: backendURL,
     })
 
-    const front = taskLambda('front', lambdaParams, { backend: backendURL })
-
-    const frontTask = new sfn.Task(scope, 'Fetch and save front', {
-        task: new tasks.InvokeFunction(front),
-    })
-    const image = taskLambda('image', lambdaParams, { backend: backendURL })
-
-    const imageTask = new sfn.Task(scope, 'Fetch Images', {
-        task: new tasks.InvokeFunction(image),
-    })
-
-    const upload = taskLambda('upload', lambdaParams)
-
-    const uploadTask = new sfn.Task(scope, 'Upload issue file', {
-        task: new tasks.InvokeFunction(upload),
+    const frontMap = new sfn.Map(scope, 'Map Fronts', {
+        inputPath: '$.',
+        itemsPath: '$.fronts',
+        parameters: {
+            'issuePublication.$': '$.issuePublication',
+            'front.$': '$$.Map.Item.Value',
+        },
+        outputPath: 'DISCARD',
     })
 
-    const zip = taskLambda('zip', lambdaParams)
-
-    const zipTask = new sfn.Task(scope, 'Make issue bundle', {
-        task: new tasks.InvokeFunction(zip),
+    const image = task(scope, 'image', 'Fetch images', lambdaParams, {
+        backend: backendURL,
     })
 
-    const indexer = taskLambda('indexer', lambdaParams)
+    const upload = task(scope, 'upload', 'Upload Issue', lambdaParams)
 
-    const indexerTask = new sfn.Task(scope, 'Generate Index', {
-        task: new tasks.InvokeFunction(indexer),
-    })
+    const zip = task(scope, 'zip', 'Make issue bundle', lambdaParams)
 
-    const notification = taskLambda('notification', lambdaParams, {
-        gu_notify_service_api_key: guNotifyServiceApiKey,
-    })
+    const indexer = task(scope, 'indexer', 'Generate Index', lambdaParams)
 
-    const notificationTask = new sfn.Task(
+    const notification = task(
         scope,
+        'notification',
         'Schedule device notification',
+        lambdaParams,
         {
-            task: new tasks.InvokeFunction(notification),
+            gu_notify_service_api_key: guNotifyServiceApiKey,
         },
     )
-    //Fetch issue metadata
-    issueTask.next(frontTask)
 
-    frontTask.next(imageTask)
+    //Fetch issue metadata
+    issue.task.next(front.task)
+
+    front.task.next(image.task)
 
     const remainingFronts = new sfn.Choice(scope, 'Check for remaining fronts')
     remainingFronts.when(
@@ -173,11 +81,11 @@ export const archiverStepFunction = (
 
     imageTask.next(remainingFronts)
 
-    uploadTask.next(zipTask)
+    upload.task.next(zip.task)
 
-    zipTask.next(indexerTask)
+    zip.task.next(indexer.task)
 
-    indexerTask.next(notificationTask)
+    indexer.task.next(notification.task)
 
     notificationTask.next(new sfn.Succeed(scope, 'successfully-archived'))
